[#1596] move runtime check for present serializer fields
